feat(navigation): move lens switcher into sidebar tabs
- Move the lens switch out of the rail into segmented tabs at the top of the sidebar, rendering only the lenses the current persona can access.
- Slim the rail to 48px (logo + footer only), reorder the footer bottom-up, and replace the Insights link with an "LFX Apps" menu (Insights / Crowdfunding / Mentorship).
- Extract the lens-aware sidebar items into SidebarNavService so the docs shell reuses the same nav: new rail + tabs, with the last active lens tab kept selected.
- Refresh the project/org selectors (border, subtle shadow, radius, plain role label) and align their dropdowns to the selector's top.
- Org lens menu: rename Overview to Dashboard, restructure sections, and update icons.
