+++ apps/lfx-one/src/app/layouts/docs-layout/docs-layout.component.ts
@@ -1,10 +1,14 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
+import { NgClass } from '@angular/common';
 import { Component, computed, CUSTOM_ELEMENTS_SCHEMA, inject } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { NavigationEnd, Router, RouterModule } from '@angular/router';
 import { LensSwitcherComponent } from '@components/lens-switcher/lens-switcher.component';
+import { SidebarComponent } from '@components/sidebar/sidebar.component';
+import { LensService } from '@services/lens.service';
+import { SidebarNavService } from '@services/sidebar-nav.service';
 import { filter, map, startWith } from 'rxjs';
 
 import { DocsSidebarNavComponent } from '../../modules/docs/components/docs-sidebar-nav/docs-sidebar-nav.component';
@@ -37,7 +41,7 @@ import { UserService } from '../../shared/services/user.service';
 @Component({
   selector: 'lfx-docs-layout',
   standalone: true,
-  imports: [RouterModule, LensSwitcherComponent, DocsSidebarNavComponent],
+  imports: [NgClass, RouterModule, LensSwitcherComponent, SidebarComponent, DocsSidebarNavComponent],
   templateUrl: './docs-layout.component.html',
   styleUrl: './docs-layout.component.scss',
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -45,8 +49,14 @@ import { UserService } from '../../shared/services/user.service';
 export class DocsLayoutComponent {
   protected readonly userService = inject(UserService);
   private readonly router = inject(Router);
+  private readonly lensService = inject(LensService);
+  private readonly sidebarNavService = inject(SidebarNavService);
 
   protected readonly isAuthenticated = computed(() => this.userService.authenticated());
+
+  // Shared lens nav — rendered faded in the docs shell so the user can return to their last lens.
+  protected readonly activeLens = this.lensService.activeLens;
+  protected readonly sidebarItems = this.sidebarNavService.sidebarItems;
 
   /**
    * `/login?returnTo=<current url>` for the mobile sign-in button. Tracks
