+++ apps/lfx-one/src/app/shared/components/org-selector/org-selector.component.html
@@ -2,8 +2,9 @@
 <!-- SPDX-License-Identifier: MIT -->
 
 <button
+  #selectorTrigger
   type="button"
-  class="box-border content-stretch flex gap-3 items-center overflow-clip p-3 relative w-full cursor-pointer hover:bg-gray-50 transition-colors rounded-lg"
+  class="box-border content-stretch flex gap-3 items-center overflow-clip p-3 relative w-full cursor-pointer hover:bg-gray-50 transition-colors rounded-[0.75rem] border border-gray-200 shadow-[0_2px_3px_-1px_rgba(0,0,0,0.08)]"
   (click)="togglePanel($event, popover)"
   [attr.aria-expanded]="popover.overlayVisible"
   data-testid="org-selector">
@@ -17,21 +18,23 @@
     </div>
   </div>
 
-  <div class="flex flex-col gap-1 grow items-start relative min-w-0 overflow-hidden">
+  <div class="flex flex-col gap-1.5 grow items-start relative min-w-0 overflow-hidden">
     <p class="sidebar-org-name" [attr.title]="displayName()" data-testid="org-selector-trigger-name">{{ displayName() }}</p>
-    @if (selectedRoleLabel()) {
-      <span
-        [attr.aria-label]="selectedRoleTooltip() ? 'Role: ' + selectedRoleLabel() + '; ' + selectedRoleTooltip() : 'Role: ' + selectedRoleLabel()"
-        class="inline-flex items-center gap-1.5 max-w-full px-2.5 py-0.5 rounded-full bg-white border border-gray-200 text-gray-700 text-xs font-medium leading-tight truncate"
-        data-testid="org-selector-trigger-role-badge"
-        [attr.data-role-persona]="selectedRolePersona()"
-        [attr.data-role-label]="selectedRoleLabel()">
-        <i [ngClass]="selectedRoleIcon()" class="text-[10px] shrink-0" aria-hidden="true"></i>
-        <span class="truncate">{{ selectedRoleLabel() }}</span>
-      </span>
-    } @else {
-      <span class="text-xs font-normal text-gray-400 leading-none truncate" data-testid="org-selector-trigger-lens-label">Organization</span>
-    }
+    <div class="flex items-center gap-1.5 w-full min-w-0 text-left">
+      @if (selectedRoleLabel()) {
+        <i [ngClass]="selectedRoleIcon()" class="fa-fw text-[10px] text-gray-400 shrink-0" aria-hidden="true"></i>
+        <span
+          class="text-xs font-normal text-gray-400 leading-none truncate"
+          data-testid="org-selector-trigger-role-badge"
+          [attr.data-role-persona]="selectedRolePersona()"
+          [attr.data-role-label]="selectedRoleLabel()"
+          [attr.title]="selectedRoleTooltip() || selectedRoleLabel()"
+          >{{ selectedRoleLabel() }}</span
+        >
+      } @else {
+        <span class="text-xs font-normal text-gray-400 leading-none truncate" data-testid="org-selector-trigger-lens-label">Organization</span>
+      }
+    </div>
   </div>
 
   <i class="fa-light fa-angles-up-down shrink-0 text-gray-400" aria-hidden="true" data-testid="org-selector-trigger-chevron"></i>
@@ -45,7 +48,7 @@
   (onShow)="onPopoverShow()"
   (onHide)="onPopoverHide()"
   data-testid="org-selector-panel">
-  <div class="p-1" data-testid="org-selector-panel-body">
+  <div data-testid="org-selector-panel-body">
     <div class="relative mb-2">
       <i class="fa-light fa-search absolute left-3 top-1/2 -translate-y-1/2 w-4 h-4 text-gray-400"></i>
       <input
@@ -108,20 +111,16 @@
                 {{ displayItem.item.name }}
               </p>
               @if (displayItem.roleLabel) {
-                <span
-                  class="inline-flex items-center gap-1.5 mt-1 max-w-full px-2 py-0.5 rounded-full bg-white border border-gray-200 text-gray-700 text-xs font-medium leading-tight truncate"
-                  [attr.aria-label]="
-                    displayItem.roleTooltip ? 'Role: ' + displayItem.roleLabel + '; ' + displayItem.roleTooltip : 'Role: ' + displayItem.roleLabel
-                  "
+                <p
+                  class="text-xs text-gray-400 leading-none mt-0.5 truncate"
                   [attr.data-testid]="'org-item-' + displayItem.item.uid + '-role-badge'"
                   [attr.data-role-label]="displayItem.roleLabel"
                   [attr.data-role-tooltip]="displayItem.roleTooltip || null"
                   [pTooltip]="displayItem.roleTooltip"
                   tooltipPosition="top"
                   [tooltipDisabled]="!displayItem.roleTooltip">
-                  <i [ngClass]="displayItem.roleIcon" class="text-[10px] shrink-0" aria-hidden="true"></i>
-                  <span class="truncate">{{ displayItem.roleLabel }}</span>
-                </span>
+                  <i [ngClass]="displayItem.roleIcon" class="mr-1" aria-hidden="true"></i>{{ displayItem.roleLabel }}
+                </p>
               }
             </div>
