+++ apps/lfx-one/src/app/shared/components/project-selector/project-selector.component.scss
@@ -19,6 +19,19 @@
     display: none !important;
   }
 
+  // stylelint-disable-next-line declaration-no-important
+  border-radius: 0.75rem !important;
+
+  // Remove PrimeNG's default panel gutter so the JS-aligned top matches the selector exactly.
+  // stylelint-disable-next-line declaration-no-important
+  margin-top: 0 !important;
+
+  // Uniform 8px padding (overrides the theme's default content padding).
+  .p-popover-content {
+    // stylelint-disable-next-line declaration-no-important
+    padding: 8px !important;
+  }
+
   // Desktop only — on mobile the sidebar is a drawer with a different layout;
   // scoping to lg+ avoids anchoring the panel off-screen on narrow viewports
   @media (min-width: 1024px) {
@@ -30,15 +43,11 @@
     // stylelint-disable-next-line declaration-no-important
     position: fixed !important;
 
-    // Force horizontal position to the right of the full sidebar (64px lens + 280px nav = 344px)
-    left: 344px !important;
+    // Position to the right of the full sidebar (48px rail + 300px nav = 348px) with an 8px gap.
+    left: 356px !important;
 
-    // Pin the popover 4px from the top of the viewport
-    top: 4px !important;
-
-    // When the impersonation banner is visible, offset by its height (48px)
-    &.project-selector-panel--with-banner {
-      top: 52px !important; // 48px banner + 4px
-    }
+    // `top` is set at runtime in onPopoverShow() to match the selector trigger's viewport
+    // position (see alignPanelTop). Keeping it out of CSS avoids drift as the layout above the
+    // selector changes (lens tabs, impersonation banner, etc.).
   }
 }
